Fundamental, public types converted to Rust.
